MNE_project/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.8.1=pypi_0
apptools=4.5.0=py_0
blas=1.0=mkl
blosc=1.17.0=h6538335_1
bz2file=0.98=py_0
bzip2=1.0.8=hfa6e2cd_1
ca-certificates=2019.9.11=hecc5488_0
cachetools=3.1.1=pypi_0
certifi=2019.9.11=py37_0
cffi=1.13.2=py37h7a1dbc1_0
chardet=3.0.4=pypi_0
configobj=5.0.6=py_0
cudatoolkit=10.1.243=h74a9793_0
curl=7.65.3=h2a8f88b_0
cycler=0.10.0=py37_0
dipy=1.0.0=py37h8aca51f_0
envisage=4.8.0=py_0
expat=2.2.5=he025d50_1004
freetype=2.9.1=ha9979f8_1
future=0.18.2=py37_0
google-auth=1.7.1=pypi_0
google-auth-oauthlib=0.4.1=pypi_0
grpcio=1.25.0=pypi_0
h5py=2.9.0=nompi_py37h9dfa0df_1103
hdf4=4.2.13=h712560f_2
hdf5=1.10.4=nompi_hcc15c50_1106
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha66f8fd_1
idna=2.8=pypi_0
imageio=2.6.1=py37_0
intel-openmp=2019.4=245
joblib=0.14.0=py_0
jpeg=9b=hb83a4c4_2
jsoncpp=1.8.4=h1ad3211_1002
kiwisolver=1.1.0=py37ha925a31_0
krb5=1.16.1=hc04afaa_7
libcurl=7.65.3=h2a8f88b_0
libiconv=1.15=hfa6e2cd_1005
libnetcdf=4.6.1=h411e497_2
libpng=1.6.37=h2a8f88b_0
libssh2=1.8.2=h642c060_2
libtiff=4.1.0=h56a325e_0
libxml2=2.9.10=h9ce36c8_0
lz4-c=1.8.3=he025d50_1001
lzo=2.10=hfa6e2cd_1000
markdown=3.1.1=pypi_0
matplotlib=3.1.1=py37hc8f65d3_0
mayavi=4.7.1=py37hf36c280_0
mkl=2019.4=245
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.0.15=py37h14836fe_0
mkl_random=1.1.0=py37h675688f_0
mne=0.19.2=py_0
mock=3.0.5=py37_0
nibabel=2.5.1=py_0
ninja=1.9.0=py37h74a9793_0
numexpr=2.7.0=py37h6538335_0
numpy=1.17.3=py37h4ceb530_0
numpy-base=1.17.3=py37hc3f5095_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py37_0
openssl=1.1.1d=hfa6e2cd_0
pandas=0.25.3=py37ha925a31_0
patsy=0.5.1=py37_0
pillow=6.2.1=py37hdc69c19_0
pip=19.3.1=py37_0
protobuf=3.10.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.7=pypi_0
pycparser=2.19=py37_0
pydicom=1.3.0=py_0
pyface=6.1.2=py_0
pygments=2.4.2=py_0
pyparsing=2.4.5=py_0
pyqt=5.9.2=py37h6538335_2
pyreadline=2.1=py37_1001
pysurfer=0.9.0=pyh24bf2e0_0
pytables=3.6.1=py37h1da0976_0
python=3.7.5=h8c8aaf0_0
python-dateutil=2.8.1=py_0
pytorch=1.3.1=py3.7_cuda101_cudnn7_0
pytz=2019.3=py_0
qt=5.9.7=vc14h73c81de_0
requests=2.22.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.0=pypi_0
scikit-learn=0.21.3=py37h6288b17_0
scipy=1.3.1=py37h29ff71c_0
seaborn=0.9.0=pyh91ea838_1
setuptools=41.6.0=py37_0
sip=4.19.8=py37h6538335_0
six=1.13.0=py37_0
sqlite=3.30.1=he774522_0
statsmodels=0.10.1=py37h8c2d366_0
tbb=2019.9=he980bc4_0
tensorboard=2.0.1=pypi_0
tk=8.6.8=hfa6e2cd_0
torchvision=0.4.2=py37_cu101
tornado=6.0.3=py37he774522_0
tqdm=4.38.0=py_0
traits=5.2.0=py37hfa6e2cd_0
traitsui=6.1.3=py_0
urllib3=1.25.7=pypi_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_0
vtk=8.2.0=py37h1e53df8_200
werkzeug=0.16.0=pypi_0
wheel=0.33.6=py37_0
wincertstore=0.2=py37_0
xz=5.2.4=h2fa13f4_4
zlib=1.2.11=h62dcd97_3
zstd=1.3.7=h508b16e_0
